// resources/js/pages/request-form/leave/types/leave.ts
export interface Leave {
    id: string;
    leave_start_date: string;
    employee_name: string;
    leave_type: string;
    leave_end_date: string;
    leave_days: string;
    status: string;
    leave_reason: string;
    leave_date_reported: string;
    leave_date_approved: string;
    leave_comments: string;
    picture: string;
    department: string;
    position: string;
    employeeid: string;
}